+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Drop support for ActiveRecord older than 6.0.
+
 ## [1.4.4] - 2022-02-07
 
 ### Fixed